+++ R/align-gg.R
@@ -57,6 +57,13 @@ align_gg <- function(mapping = aes(), size = NULL, action = NULL,
                      plot_data = deprecated(), theme = deprecated(),
                      free_labs = deprecated()) {
     assert_mapping(mapping)
+    action <- check_action(action, TRUE)
+    action <- deprecate_action(
+        action, "align_gg",
+        plot_data, theme,
+        free_spaces, free_labs,
+        free_guides = free_guides
+    )
     align(AlignGG,
         params = list(mapping = mapping),
         size = size, data = data %||% waiver(), action = action,