+++ R/layout-heatmap-oncoplot.R
@@ -76,7 +76,7 @@ ggoncoplot.functon <- ggoncoplot.NULL
 #' @export
 ggoncoplot.formula <- ggoncoplot.functon
 
-#' @importFrom vctrs vec_slice
+#' @importFrom vctrs vec_slice list_sizes vec_rep_each
 #' @importFrom ggplot2 aes
 #' @export
 ggoncoplot.default <- function(data, mapping = aes(), ...,
@@ -147,20 +147,10 @@ ggoncoplot.default <- function(data, mapping = aes(), ...,
             hmanno("t") +
             align_order(order(column_scores, decreasing = TRUE))
     }
-    ans +
-        hmanno(NULL,
-            action = plot_action(data = function(data) {
-                separate_longer_alt(data)
-            })
-        ) +
-        # we always set the default plot action data for the heatmap body
-        plot_action(data = function(data) separate_longer_alt(data))
-}
-
-#' @importFrom vctrs list_sizes vec_rep_each
-separate_longer_alt <- function(data, delim = "[;:,|]") {
-    value_list <- strsplit(data$value, split = delim)
-    data <- vec_rep_each(data, list_sizes(value_list))
-    data$value <- unlist(value_list, recursive = FALSE, use.names = FALSE)
-    data
+    ans + hmanno(NULL, action = plot_action(data = function(data) {
+        value_list <- strsplit(data$value, split = "[;:,|]")
+        data <- vec_rep_each(data, list_sizes(value_list))
+        data$value <- unlist(value_list, recursive = FALSE, use.names = FALSE)
+        data
+    }))
 }